+++ src/GammonX/GammonX.Mars.Server.Tests/Controller/EvalControllerTests.cs
@@ -298,14 +298,8 @@ namespace GammonX.Mars.Server.Tests.Controller
             Assert.NotNull(cubeEval);
             Assert.Equal("OK", cubeEval.Type);
             Assert.IsType<CubeEvalPayload>(cubeEval.Payload);
-            if (!isWhite)
-            {
-                Assert.Equal(CubeAction.NoDouble, cubeEval.Payload.ShouldOffer);
-            }
-            else
-            {
-                Assert.Equal(CubeAction.NoDouble, cubeEval.Payload.ShouldOffer);
-            }
+            Assert.Equal(CubeAction.TooGood, cubeEval.Payload.ShouldOffer);
+            Assert.Equal(CubeAction.Take, cubeEval.Payload.ShouldTake);
         }
 
         [Theory]