+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 18.08.2026
+
+### NEW
+- cube actions as history events
+
+### FIXES
+- fixed cube eval when using nn model
+- fixed some recover roll quirks in the implementation
+- some fixes in cube eval on edge cases (e.g. re-double)
+- fixed backgammon board contract deserialization
+
 ## 14.06.2026
 
 ### NEW